+++ packages/Datadog.Unity/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## Unreleased
+
+* Fixed a bug that could cause a NullReferenceException when a `DatadogTrackedWebRequest` was canceled.
+
 ## 1.4.0
 
 * Fix compatibility issues (crash) with Unity 6.